% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CacheConnection-class.R
\name{CacheConnection}
\alias{CacheConnection}
\title{CacheConnection R6 Class}
\description{
An R6 class for handling RDS caching for data analysis and report generation.
}
\keyword{CacheConnection}
\keyword{Initialize}
\keyword{class}
\keyword{internal}
\keyword{the}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CacheConnection-new}{\code{CacheConnection$new()}}
\item \href{#method-CacheConnection-load_from_disk}{\code{CacheConnection$load_from_disk()}}
\item \href{#method-CacheConnection-save_to_disk}{\code{CacheConnection$save_to_disk()}}
\item \href{#method-CacheConnection-append_page_note}{\code{CacheConnection$append_page_note()}}
\item \href{#method-CacheConnection-get_notes}{\code{CacheConnection$get_notes()}}
\item \href{#method-CacheConnection-reactive}{\code{CacheConnection$reactive()}}
\item \href{#method-CacheConnection-set_cache_path}{\code{CacheConnection$set_cache_path()}}
\item \href{#method-CacheConnection-set_national_estimates}{\code{CacheConnection$set_national_estimates()}}
\item \href{#method-CacheConnection-set_regional_estimates}{\code{CacheConnection$set_regional_estimates()}}
\item \href{#method-CacheConnection-set_quality_data}{\code{CacheConnection$set_quality_data()}}
\item \href{#method-CacheConnection-clone}{\code{CacheConnection$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheConnection-new"></a>}}
\if{latex}{\out{\hypertarget{method-CacheConnection-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheConnection$new(rds_path = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rds_path}}{Path to the RDS file (NULL for in-memory only).
Load data from RDS file}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheConnection-load_from_disk"></a>}}
\if{latex}{\out{\hypertarget{method-CacheConnection-load_from_disk}{}}}
\subsection{Method \code{load_from_disk()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheConnection$load_from_disk()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
None
Save data to disk (only if changed and RDS path is not NULL)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheConnection-save_to_disk"></a>}}
\if{latex}{\out{\hypertarget{method-CacheConnection-save_to_disk}{}}}
\subsection{Method \code{save_to_disk()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheConnection$save_to_disk()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
None
Reactive method for the CacheConnection object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheConnection-append_page_note"></a>}}
\if{latex}{\out{\hypertarget{method-CacheConnection-append_page_note}{}}}
\subsection{Method \code{append_page_note()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheConnection$append_page_note(
  page_id,
  object_id,
  note,
  parameters = list(),
  include_in_report = FALSE,
  include_plot_table = FALSE,
  single_entry = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheConnection-get_notes"></a>}}
\if{latex}{\out{\hypertarget{method-CacheConnection-get_notes}{}}}
\subsection{Method \code{get_notes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheConnection$get_notes(page_id, object_id = NULL, parameters = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheConnection-reactive"></a>}}
\if{latex}{\out{\hypertarget{method-CacheConnection-reactive}{}}}
\subsection{Method \code{reactive()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheConnection$reactive()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A Shiny reactive expression tracking changes
Update a field (with change tracking)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheConnection-set_cache_path"></a>}}
\if{latex}{\out{\hypertarget{method-CacheConnection-set_cache_path}{}}}
\subsection{Method \code{set_cache_path()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheConnection$set_cache_path(value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheConnection-set_national_estimates"></a>}}
\if{latex}{\out{\hypertarget{method-CacheConnection-set_national_estimates}{}}}
\subsection{Method \code{set_national_estimates()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheConnection$set_national_estimates(value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheConnection-set_regional_estimates"></a>}}
\if{latex}{\out{\hypertarget{method-CacheConnection-set_regional_estimates}{}}}
\subsection{Method \code{set_regional_estimates()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheConnection$set_regional_estimates(value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheConnection-set_quality_data"></a>}}
\if{latex}{\out{\hypertarget{method-CacheConnection-set_quality_data}{}}}
\subsection{Method \code{set_quality_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheConnection$set_quality_data(value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheConnection-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CacheConnection-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheConnection$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
