% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cd2030.pooled-package.R
\docType{package}
\name{cd2030.pooled-package}
\alias{cd2030.pooled}
\alias{cd2030.pooled-package}
\title{cd2030.pooled: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: David Kariuki \email{hello@damurka.com} (\href{https://orcid.org/0009-0003-6159-1107}{ORCID})

Authors:
\itemize{
  \item Martin Mutua
  \item Peter Kaberia \email{pkaberia@aphrc.org}
  \item Godfrey Adero \email{gadero@aphrc.org}
  \item Rornald Muhumuza \email{rmuhumuza@aphrc.org}
}

Other contributors:
\itemize{
  \item African Population and Health Research Center (APHRC) [copyright holder, funder]
}

}
\keyword{internal}
